+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Keep it human-readable, your future self will thank you!
 
 ## [0.3.1 - AIFS v0.3 Compatibility](https://github.com/ecmwf/anemoi-training/compare/0.3.0...0.3.1) - 2024-11-28
 
+### Changed
+- Perform full shuffle of training dataset [#153](https://github.com/ecmwf/anemoi-training/pull/153)
+
 ### Fixed
 
 - Fixed bug in power spectra plotting for the n320 resolution. [#149](https://github.com/ecmwf/anemoi-training/pull/149)